Läuft jetzt auch für double Zeiten

+++ Blatt-03/Aufgabe1.h
@@ -93,7 +93,7 @@ void Runge_Kutta( double Zeit , double *Ort , double *Geschwindigkeit , double* 
 
 double* Runge_Kutta( double Zeit , double *Ort , double *Geschwindigkeit , double* Kraftfeld(double*) , double Breite = 0.001){
 	double *Y_n = new double[7];
-	int Anzahl = (int) Zeit / Breite;
+	double Anzahl = Zeit / Breite;
 	for (int i = 0; i < 3; ++i)
 	{
 		Y_n[i] = Ort[i];
@@ -107,7 +107,7 @@ double* Runge_Kutta( double Zeit , double *Ort , double *Geschwindigkeit , doubl
 	double *K_3=new double[6];
 	double *K_4=new double[6];
 	
-	for (int Zaehler = 1; Zaehler <= Anzahl; ++Zaehler)
+	for (double Zaehler = 1; Zaehler <= Anzahl; ++Zaehler)
 	{
 		for (int Komponente = 0; Komponente < 3; ++Komponente)
 		{
